Refactor TOTAL_MATERIAL and others

+++ src/Pawn.cpp
@@ -3,15 +3,15 @@
 
 #include "Pawn.hpp"
 
-unsigned int Pawn::pawnScore(const Board &board, const int64_t gameStage, const int64_t alpha)
+unsigned int Pawn::pawnScore(const Board &board, const int gameStage, const int alpha)
 {
     unsigned int doubled = doubledPawns(board);
     unsigned int passed = passedPawns(board);
     unsigned int isolated = isolatedPawns(board);
 	return (passed*(EvalTables::PawnTable[OPENING][PASSED]*gameStage + EvalTables::PawnTable[ENDGAME][PASSED]*alpha) +
 	doubled*(EvalTables::PawnTable[OPENING][DOUBLED]*gameStage + EvalTables::PawnTable[ENDGAME][DOUBLED]*alpha) +
-	isolated*(EvalTables::PawnTable[OPENING][ISOLATED]*gameStage + EvalTables::PawnTable[ENDGAME][ISOLATED]*alpha))
-	/Eval::TOTAL_MATERIAL;
+	isolated*(EvalTables::PawnTable[OPENING][ISOLATED]*gameStage + EvalTables::PawnTable[ENDGAME][ISOLATED]*alpha))/
+	Eval::TOTAL_MATERIAL;
 }
 
 unsigned int Pawn::doubledPawns(const Board &board)
